@@ -7,7 +7,7 @@ version: 0.0.1+1
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0"
+  sdk: '>=2.12.0 <4.0.0'
 
 dependencies:
   flutter:
